requirements.txt
azure-identity==1.12.0
aiohttp==3.8.4
tenacity==8.2.2
python-dotenv==0.21.1
tabulate==0.9.0
